Refactor partial_tempering to make it testable
